requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==3.6.4.0
certifi==2024.2.2
chardet==4.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crayons==0.4.0
diff-match-patch==20230430
distlib==0.3.8
Django==4.1.13
django-import-export==4.0.1
django-model-utils==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
fffw==4.0.0
filelock==3.14.0
idna==2.10
kombu==5.2.4
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
pymediainfo==5.1.0
pyotp==2.9.0
pytz==2024.1
PyYAML==6.0.1
requests==2.25.1
sqlparse==0.5.0
tablib==3.5.0
urllib3==1.26.18
vine==5.1.0
virtualenv==20.26.1
wcwidth==0.2.13
